+++ gds2x/povray/gdsobject_pov.cpp
@@ -66,7 +66,7 @@ void GDSObject_pov::OutputPaths()
 			if(path->GetWidth()){
 				fprintf(m_optr, "mesh2 { vertex_vectors { %d", 4*(path->GetPoints()));
 
-				float x, y, z;
+				double x, y, z;
 				for(unsigned int idx=0; path->GetPoint3D(idx, x, y, z); idx++){
 					fprintf(m_optr, ",<%.2f,%.2f,%.2f>", x, y, -z);
 				}
@@ -133,7 +133,7 @@ void GDSObject_pov::OutputTexts()
 				if(text->GetRY()){
 					fprintf(m_optr, "Rotate_Around_Trans(<0,0,%.2f>,<%.2f,%.2f,%.2f>)", -text->GetRY(), text->GetX(), text->GetY(), -text->GetZ());
 				}
-				float htrans = 0.0, vtrans = 0.0;
+				double htrans = 0.0, vtrans = 0.0;
 				switch(text->GetHJust()){
 					case 0:
 						htrans = -0.5*text->GetString().length();
@@ -194,7 +194,7 @@ void GDSObject_pov::OutputARefs()
 	for(unsigned int i = 0; i < m_arefs.size(); i++){
 		GDS2X::ARefElement *aref = m_arefs[i];
 
-		float dx = 0.0, dy = 0.0;
+		double dx = 0.0, dy = 0.0;
 		int columns = 0, rows = 0;
 
 		switch((int)round(aref->rotate.y)){
@@ -202,8 +202,8 @@ void GDSObject_pov::OutputARefs()
 			case 360:
 			case 180:
 			case -180:
-				dx = (float)(aref->x2 - aref->x1) / (float)aref->columns;
-				dy = (float)(aref->y3 - aref->y1) / (float)aref->rows;
+				dx = (aref->x2 - aref->x1) / aref->columns;
+				dy = (aref->y3 - aref->y1) / aref->rows;
 				columns = aref->columns;
 				rows = aref->rows;
 				break;
@@ -212,8 +212,8 @@ void GDSObject_pov::OutputARefs()
 			case -270:
 			case 270:
 			case -90:
-				dx = (float)(aref->x3 - aref->x1) / (float)aref->rows;
-				dy = (float)(aref->y2 - aref->y1) / (float)aref->columns;
+				dx = (aref->x3 - aref->x1) / aref->rows;
+				dy = (aref->y2 - aref->y1) / aref->columns;
 				columns = aref->rows;
 				rows = aref->columns;
 				break;
